requirements.txt
torch
opencv-python
pandas
mediapipe
deepface
matplotlib
transformers
torchvision
Pillow
python-dotenv
